Remove trailing whitespace
As uncritical as it is, but there is definitely a limit of how much
trailing whitespace can be tolerated.

I tried not to touch files where trailing whitespace could make a
difference (or makes no difference whatsoever). Brave souls are
encouraged to review the diff.

A more crucial thing would be to have no mixed identation (in this
particular case tabs have to be replaced with spaces). This is
especially important in .py files, and I'm even surprised that things
are working correctly right now. A simple approach of changing all
tabs to 4 spaces is not going to work given that some cases will have
to be dealt with manually… but that's too much of work for what it's
worth, at least until issue #12 is resolved.

As a (positive) side effect of stripping out trailing whitespace, in a
couple of files CR-LF line endings were changed to LF.