Basic Verilog files
This commit adds some files translated to (System)Verilog. Note that
verilog files existed in the repo already, some of them were
rewritten, others were simply deleted. Deleted files may be re-added
later.

The code is written in a way to be parseable by tools that may not
support all SystemVerilog features, but still using some of the
SystemVerilog features that are typically available. Some of these may
be downgraded later for compatibility.

Tests for these files will be committed separately.